// lib/types.ts
export type SipMethod = 'ACK' | 'BYE' | 'CANCEL' | 'INFO' | 'INVITE' | 'MESSAGE' | 'NOTIFY' | 'OPTIONS' | 'PRACK' | 'PUBLISH' | 'REFER' | 'REGISTER' | 'SUBSCRIBE' | 'UPDATE';
export type SipMessageHeaders = Record<string, string>;
export type AOR = { name: string; uri: string; params?: Record<string, any>; };
export type Via = { version: string; protocol: string; host: string; port: string; };

export interface SrfConfig {
  apiSecret?: string;
  host?: string;
  port?: number;
  secret?: string;
  tags?: string[];
}

export interface ParseUriResult {
  family?: 'ipv6' | 'ipv4';
  scheme: 'sip' | 'sips' | 'tel';
  user?: string;
  password?: string;
  host?: string;
  port?: string;
  params: Record<string, string | null>;
  headers: Record<string, string>;
  context?: string;
}

export interface SipMessage {
  type: "request" | "response";
  body: string;
  payload: object[];
  source: "network" | "application";
  source_address: string;
  source_port: string;
  protocol: string;
  stackTime: string;
  calledNumber: string;
  callingNumber: string;
  raw: string;
  get(name: string): string;
  has(name: string): boolean;
  set(name: string, value: string): void;
  getParsedHeader(name: "contact" | "Contact"): Array<AOR>;
  getParsedHeader(name: "via" | "Via"): Array<Via>;
  getParsedHeader(name: "To" | "to" | "From" | "from" | "refer-to" | "referred-by" | "p-asserted-identity" | "remote-party-id"): AOR;
  getParsedHeader(name: string): any;
}

export interface SrfRequest extends SipMessage {
  method: SipMethod;
  readonly isNewInvite: boolean;
  stackDialogId?: string;
  canceled: boolean;
  _dialogState: any;

  cancel(opts?: { headers?: Record<string, string> }, callback?: (err: any, req: SrfRequest) => void): void;
  on(event: 'response', callback: (res: SrfResponse, ack: (opts?: { body: string }) => void) => void): this;
  on(event: 'cancel', callback: (res: SrfRequest) => void): this;
  on(event: 'update', callback: (req: SrfRequest, res: SrfResponse) => void): this;
  proxy(opts: ProxyRequestOptions, callback: (err: any, results: any) => void): void;
  listeners(event: string): Function[];
  emit(event: string, ...args: any[]): boolean;
  auth: any;
  stackTxnId: string;

  branch: string;
  callId: string;
  from: string;
  headers: Record<string, string>;
  msg: any;
  sdp: string;
  srf: any;
  to: string;
  uri: string;
  registration?: {
    type: "unregister" | "register";
    expires: number;
    contact: Array<AOR>;
    aor: string;
  };
  getHeaderName(hdr: string): string | undefined;
  get(name: string): string;
  socket: Socket;
  callingName?: string;
}

export interface ProxyRequestOptions {
  forking?: 'sequential' | 'simultaneous';
  remainInDialog?: boolean;
  recordRoute?: boolean;
  provisionalTimeout?: string;
  finalTimeout?: string;
  followRedirects?: boolean;
  destination?: string | string[];
}

export interface SrfResponse extends SipMessage {
  srf: any;
  headers: SipMessageHeaders;
  status: number;
  statusCode: number;
  reason: string;
  finalResponseSent: boolean;
  req: SrfRequest;
  agent: any;
  socket: any;
  send(status: number, opts?: object, cb?: () => void): void;
  send(status: number, reason?: string, opts?: object, cb?: () => void): void;
  send(status: number, reason?: string, opts?: object, cb?: (err: any, msg: SipMessage) => void): void;
  end(): void;
  get(name: string): string;
  getHeaderName(hdr: string): string | undefined;
}